# packages/schemas/openapi/gridworks-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GRIDWORKS API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080"
    }
  ],
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 160
        }
      },
      "PlayerID": {
        "name": "player_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "CompanyID": {
        "name": "company_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Stable error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "timestamp"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            },
            "additionalProperties": false
          }
        }
      },
      "CompanyCreate": {
        "type": "object",
        "required": [
          "company_type",
          "name"
        ],
        "properties": {
          "company_type": {
            "enum": [
              "holding",
              "operating"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 80
          }
        },
        "additionalProperties": false
      },
      "GroupCreate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 80
          }
        },
        "additionalProperties": false
      },
      "ProfilePatch": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "bio": {
            "type": "string",
            "maxLength": 1000
          },
          "locale": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "visibility": {
            "enum": [
              "public",
              "private"
            ]
          },
          "dm_policy": {
            "enum": [
              "everyone",
              "nobody"
            ]
          },
          "discoverable": {
            "type": "boolean"
          },
          "notification_preferences": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        },
        "additionalProperties": false
      }
    }
  },
  "paths": {
    "/healthz": {
      "get": {
        "responses": {
          "200": {
            "description": "Healthy"
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "responses": {
          "200": {
            "description": "Ready"
          },
          "503": {
            "description": "Persistence unavailable"
          }
        }
      }
    },
    "/version": {
      "get": {
        "responses": {
          "200": {
            "description": "Version metadata"
          }
        }
      }
    },
    "/api/v1/auth/guest": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Guest identity or safe replay"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated self"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/me/profile": {
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/players/{player_id}": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/PlayerID"
          }
        ],
        "responses": {
          "200": {
            "description": "Public player"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/companies": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created company"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/CompanyID"
          }
        ],
        "responses": {
          "200": {
            "description": "Public company"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/company-groups": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created group"
          }
        }
      }
    }
  }
}